# docs/swagger.yaml
definitions:
  collections.createCollectionRequest:
    properties:
      name:
        type: string
    type: object
  collections.createCollectionResponse:
    properties:
      id:
        type: string
    type: object
  collections.updateCollectionRequest:
    properties:
      name:
        type: string
    type: object
  domain.CollectionDB:
    properties:
      createdAt:
        type: integer
      id:
        type: string
      name:
        type: string
      updatedAt:
        type: integer
    type: object
  domain.GalleryDB:
    properties:
      collectionId:
        type: string
      createdAt:
        type: integer
      id:
        type: string
      name:
        type: string
      passwordAttempts:
        type: integer
      passwordHash:
        type: string
      sharingEnabled:
        type: boolean
      sharingExpiryDate:
        type: integer
      updatedAt:
        type: integer
    type: object
  fiber.Map:
    additionalProperties: true
    type: object
  galleries.createGalleryRequest:
    properties:
      name:
        example: Example Gallery
        type: string
    type: object
  galleries.createGalleryResponse:
    properties:
      id:
        type: string
    type: object
  galleries.generateQrRequest:
    description: Request body for generating a QR code
    properties:
      url:
        description: URL to be encoded in the QR code
        example: https://example.com
        type: string
    type: object
info:
  contact:
    name: Michał Klemens
  description: Image galleries that provides uploading images
  title: Studio Ginger - QR code generator
  version: "0.1"
paths:
  /api/v1/collections:
    get:
      consumes:
      - '*/*'
      description: Gets all collections
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.CollectionDB'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/fiber.Map'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Get collections
      tags:
      - collections
    post:
      consumes:
      - application/json
      description: Creates one collection
      parameters:
      - description: Collection to create
        in: body
        name: collections
        required: true
        schema:
          $ref: '#/definitions/collections.createCollectionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/collections.createCollectionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fiber.Map'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/fiber.Map'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Create collection
      tags:
      - collections
  /api/v1/collections/{collectionId}:
    delete:
      consumes:
      - '*/*'
      description: Deletes specific collection
      parameters:
      - description: Collection ID
        in: path
        name: collectionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/fiber.Map'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Delete collection
      tags:
      - collections
    get:
      consumes:
      - '*/*'
      description: Gets specific collection
      parameters:
      - description: Collection ID
        in: path
        name: collectionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.CollectionDB'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/fiber.Map'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Get collection
      tags:
      - collections
    put:
      consumes:
      - '*/*'
      description: Updates specific collection
      parameters:
      - description: Collection update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/collections.updateCollectionRequest'
      - description: Collection ID
        in: path
        name: collectionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.CollectionDB'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fiber.Map'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/fiber.Map'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/fiber.Map'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Update collection
      tags:
      - collections
  /api/v1/collections/{collectionId}/galleries:
    get:
      consumes:
      - '*/*'
      description: gets all galleries of a collection with collectionId.
      parameters:
      - description: Collection ID
        in: path
        name: collectionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.GalleryDB'
            type: array
      summary: Get all galleries of a collection.
      tags:
      - collections
    post:
      consumes:
      - application/json
      description: Creates one gallery in collection with collectionId
      parameters:
      - description: Gallery to create
        in: body
        name: galleries
        required: true
        schema:
          $ref: '#/definitions/galleries.createGalleryRequest'
      - description: Collection ID
        in: path
        name: collectionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/galleries.createGalleryResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/fiber.Map'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/fiber.Map'
      summary: Create one gallery
      tags:
      - collections
  /api/v1/galleries/{galleryId}/qr:
    post:
      consumes:
      - application/json
      description: Generate a QR code from a given URL
      parameters:
      - description: QR Generation Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/galleries.generateQrRequest'
      - description: Collection ID
        in: path
        name: collectionId
        required: true
        type: string
      - description: Gallery ID
        in: path
        name: galleryId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Generate QR code
      tags:
      - QR
swagger: "2.0"